--- a/app/jobs/destroy_problems_by_id_job.rb
+++ b/app/jobs/destroy_problems_by_id_job.rb
@@ -3,6 +3,7 @@
 class DestroyProblemsByIdJob < ActiveJob::Base
   queue_as :default
 
+  # @param problem_ids [Array<String>]
   def perform(problem_ids)
     bson_problem_ids = []