Before:

```
rspec spec/jobs/destroy_problems_by_id_job_spec.rb --profile

Randomized with seed 62023

DestroyProblemsByIdJob
  destroys all selected problems, even with a large amount of them
  should work with a fresh new application
  destroys all selected problems

Top 3 slowest examples (15.75 seconds, 100.0% of total time):
  DestroyProblemsByIdJob destroys all selected problems, even with a large amount of them
    15.57 seconds ./spec/jobs/destroy_problems_by_id_job_spec.rb:24
  DestroyProblemsByIdJob should work with a fresh new application
    0.10749 seconds ./spec/jobs/destroy_problems_by_id_job_spec.rb:38
  DestroyProblemsByIdJob destroys all selected problems
    0.07038 seconds ./spec/jobs/destroy_problems_by_id_job_spec.rb:16

Finished in 15.75 seconds (files took 1.93 seconds to load)
3 examples, 0 failures
```

After:

```
rspec spec/jobs/destroy_problems_by_id_job_spec.rb --profile

Randomized with seed 12988

DestroyProblemsByIdJob
  destroys all selected problems
  should work with a fresh new application

Top 2 slowest examples (0.28319 seconds, 99.1% of total time):
  DestroyProblemsByIdJob destroys all selected problems
    0.25373 seconds ./spec/jobs/destroy_problems_by_id_job_spec.rb:6
  DestroyProblemsByIdJob should work with a fresh new application
    0.02946 seconds ./spec/jobs/destroy_problems_by_id_job_spec.rb:17

Finished in 0.28575 seconds (files took 2.34 seconds to load)
2 examples, 0 failures
```

Total: minus 15 seconds